refactor: Tags -> TagList
