+++ src/PostListItem.ts
@@ -1,5 +1,5 @@
 import { SummaryEntity, generateSummaryEntity } from 'search-summary';
-import { TagList } from './Tags';
+import { TagList } from './TagList';
 import { formatYMD, fromNow } from './lib/date-utils';
 import { Post } from './post';
