+++ src/css/elements.css
@@ -36,13 +36,13 @@
 /**
  * 記事タグ
  */
-.bl_tags {
+.bl_tagList {
   display: inline-flex;  /* like `span`, but flexbox inside */
   flex-wrap: wrap;
   list-style: none;
 }
 
-.bl_tags li {
+.bl_tagList li {
   margin-left: 0.5em;
 }
